replaced int with long

+++ src/main/java/ru/netology/stats/StatsService.java
@@ -1,7 +1,7 @@
 package ru.netology.stats;
 
 public class StatsService {
-    public long minSales(long[] sales) {
+    public int minSales(int[] sales) {
         int minMonth = 0; // номер месяца с минимальными продажами среди просмотренных ранее
 
         for (int i = 0; i < sales.length; i++) {
@@ -13,7 +13,7 @@ public class StatsService {
         return minMonth + 1; // месяца нумеруются с 1, а индексы массива с 0, нужно сдвинуть ответ на 1
     }
 
-    public long maxSales(long[] sales) {
+    public int maxSales(int[] sales) {
         int maxMonth = 0; //номер месяца с максимальной продажой
 
         for (int t = 0; t < sales.length; t++) {
@@ -43,9 +43,9 @@ public class StatsService {
         return amount / sales.length;
     }
 
-    public long belowAverageSales(long[] sales) {
+    public int belowAverageSales(long[] sales) {
         long averge = averageAmount(sales);
-      long count = 0;
+      int count = 0;
 
         for (long sale : sales) {
             if (sale <= averge) {
@@ -55,9 +55,9 @@ public class StatsService {
         return count;
     }
 
-    public long aboveAverageSales(long[] sales) {
+    public int aboveAverageSales(long[] sales) {
         long averge = averageAmount(sales);
-        long count = 0;
+        int count = 0;
 
         for (long sale : sales) {
             if (sale > averge) {
